requirements.txt

blinker==1.7.0
charset-normalizer==3.3.2
comtypes==1.2.0
contourpy==1.1.1
cryptography==43.0.3
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
et-xmlfile==2.0.0
ezdxf==1.1.2
filelock==3.13.1
flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.10.0
itsdangerous==2.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
openpyxl==3.1.5
packaging==23.2
pdf2docx==0.5.8
pdf2image==1.16.3
pdf2pptx==1.0.5
pdfminer.six==20231228
pdfplumber==0.11.4
pdfrw==0.4
Pillow==10.1.0
ppt2pdf==0.0.3





PyInquirer==1.0.3
PyMuPDF==1.20.1
PyMuPDFb==1.23.6

pypdf2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.8.2
python-docx==1.1.2
python-pptx==0.6.23


reportlab==4.0.7
requests==2.31.0
six==1.16.0


sympy==1.12

tinycss2==1.2.1
tomli==2.0.1



typing-extensions==4.12.2
tzdata==2024.2
unoconv==0.9.0
urllib3==2.1.0
wcwidth==0.2.12
webencodings==0.5.1
werkzeug==3.0.1
XlsxWriter==3.1.9
zipp==3.17.0
